cloud_run_predict/requirements.txt
absl-py==1.0.0
astroid==2.9.3
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
black==21.12b0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.1
colorama==0.4.4
coverage==6.3.1
cvxpy==1.2.1
cycler==0.10.0
Cython==0.29.24
decorator==5.1.1
ecos==2.0.10
exchange-calendars==3.6.2
executing==0.8.2
Flask==2.0.2
Flask-WTF==0.15.1
flatbuffers==1.12
gast==0.4.0
google-api-core==2.5.0
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-cloud-aiplatform==1.13.0
google-cloud-bigquery==2.32.0
google-cloud-core==2.3.0
google-cloud-resource-manager==1.4.1
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.1
grpc-google-iam-v1==0.12.4
grpcio==1.43.0
grpcio-status==1.43.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
ipython==8.0.0
isort==5.10.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.0.1
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
korean-lunar-calendar==0.2.1
lazy-object-proxy==1.7.1
libclang==14.0.1
lxml==4.8.0
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
multitasking==0.0.10
mypy-extensions==0.4.3
numpy==1.21.2
oauthlib==3.2.0
opt-einsum==3.3.0
osqp==0.6.2.post5
packaging==21.3
pandas==1.3.3
pandas-market-calendars==3.4
parso==0.8.3
pathspec==0.9.0
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
platformdirs==2.4.1
pluggy==1.0.0
prompt-toolkit==3.0.26
proto-plus==1.19.9
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.11.2
pylint==2.12.2
pyluach==1.4.1
pyparsing==2.4.7
pyportfolioopt==1.5.2
pytest==6.2.5
pytest-cov==3.0.0
python-dateutil==2.8.2
pytz==2021.3
qdldl==0.1.5.post2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
scs==3.2.0
six==1.16.0
sklearn==0.0
stack-data==0.1.4
statsmodels==0.13.1
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.0
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.3
toolz==0.11.2
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.13.3
WTForms==2.3.3
yfinance==0.1.70
zipp==3.8.0
